Update README.md
right now i'm learning how to use branches. from my understanding branches is like saving another copy for you to tinker on. This is separate from the main branch. the main branch is the final draft. however, if you want to do some edits to it, it's better to make a new copy of our "final" draft and edit on a new copy. Then, after we read over the new version, we can see if we want to make any changes into our main branch.